fixed some bugs and moved a couple of style sheets around
