+++ src/screens/CoffeeChat/CoffeeChat.css
@@ -1,0 +1,171 @@
+
+header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 30px;
+}
+h1 {
+    margin: 40px;
+    font-family: 'Itim';
+    font-size: 60px;
+    font-weight: 600;
+    color: #66451C;
+}
+
+.search-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    background-color: white;
+    max-width: 800px;
+    width: 100%;
+    border: 3px solid #8b4513;
+    border-radius: 20px;
+    margin: 30px 20px;
+    padding: 5px;
+}
+
+.search-input {
+    width: 800px;
+    margin: 5px 10px;
+    border: none;
+    font-size: medium;
+}
+
+.search-input:focus {
+    outline: none;
+    color: #142b0d;
+  }
+
+.categories {
+    display: flex;
+    gap: 15px;
+    margin-bottom: 20px;
+    overflow-x: auto;
+    padding-bottom: 10px;
+    max-width: 800px;
+    width: 100%;
+}
+
+.category {
+    background-color: #D0E1D1;
+    border: 1px solid #10300E;
+    color: #10300E;
+    padding: 8px 16px;
+    border-radius: 20px;
+    font-size: 14px;
+    font-weight: 600;
+    cursor: pointer;
+    white-space: nowrap;
+}
+
+.category.active {
+    background-color: #3B6838;
+    color: #fff;
+}
+
+.students-list {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 20px;
+    margin-bottom: 30px;
+}
+
+.student-card {
+    background-color: #D0E1D1;
+    border-radius: 25px;
+    overflow: hidden;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+    display: flex;
+    flex-direction: column;
+}
+
+.student-card:hover {
+    transform: translateY(-3px);
+}
+
+.student-topbar {
+    background-color: #3B6838;
+    color: #fff;
+    padding: 20px;
+    width: 100%;
+    flex-shrink: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+}
+
+.student-name {
+    font-size: 18px;
+    font-weight: 600;
+    margin-bottom: 8px;
+}
+
+.student-major {
+    display: inline-block;
+    background-color: #fff;
+    color: #000000;
+    padding: 5px 10px;
+    border-radius: 15px;
+    font-size: 14px;
+    margin-top: 5px;
+    width: fit-content;
+}
+
+.student-info {
+    padding: 20px;
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    flex-grow: 1;
+    align-items: center;
+    text-align: center;
+}
+
+.student-bio {
+    color: #000000;
+    font-size: 14px;
+    line-height: 1.5;
+}
+
+.connect-button {
+    background-color: #E8FABF;
+    color: #000000;
+    border: none;
+    border-radius: 25px;
+    padding: 12px 20px;
+    font-size: 15px;
+    font-weight: 600;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 8px;
+    margin-top: auto;
+    width: 140px;
+}
+
+.connect-button:hover {
+    background-color: #3B6838;
+}
+
+@media (max-width: 768px) {
+    .student-card {
+    flex-direction: column;
+}
+
+.student-sidebar {
+    width: 100%;
+    padding: 15px;
+}
+
+.connect-button {
+    width: 80%;
+    }
+}
