+++ src/screens/CoffeeChat/RegisterCoffeeChat.css
@@ -8,13 +8,7 @@ body {
     justify-content: center;
     min-height: 100vh;
   }
-  
-  .container {
-    max-width: 600px;
-    width: 100%;
-    padding: 20px;
-  }
-  
+
   .page-title-title {
     text-align: center;
     font-size: 24px;
