+++ src/screens/LandingPage/LandingPage.css
@@ -1,58 +1,39 @@
-@import url('https://fonts.googleapis.com/css2?family=Itim&display=swap');
-
-.itim-regular {
-  font-family: "Itim", cursive;
-  font-weight: 200;
-  font-style: normal;
-}
-.body {
-  margin: 0;
-  padding: 0;
-  font-family: Helvetica, Sans-Serif;
-}
-
-h1, h2, h3 {
-  font-family: 'Itim';
-}
-
-p { 
-  font-family: Arial, Helvetica, sans-serif;
-}
-
 .container {
   display: flex;
   flex-direction: column;
   align-items: center;
   background-color: #edece4;
+  height: 100vh;
+  width: 100vw;
 }
 
 .navbar {
   position: relative;
   z-index: 1000;
-  width: 100%;
+  width: 100vw;
   background-color: #3b6838;
   color: white;
   display: flex;
-  align-items: center; 
+  align-items: center;
   padding: 15px 30px;
 }
 
 .title {
-  margin: 0 auto;  
+  margin: 0 auto;
   font-size: 33px;
-  text-align: center;  
-  position: relative;  
+  text-align: center;
+  position: relative;
 }
 
 .hamburger {
-  cursor: pointer; 
-  margin: 5px 30px;  
+  cursor: pointer;
+  margin: 5px 30px;
   padding: 7px;
   border: 3px solid white;
   border-radius: 50%;
   position: absolute;
-  left: 15px; 
-  z-index: 1100; 
+  left: 15px;
+  z-index: 1100;
 }
 
 .savedCourses {
@@ -73,7 +54,6 @@ p {
 
 .body {
   text-align: center;
-  margin: 20px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -83,6 +63,7 @@ p {
 h2 {
   color: #3b6838;
   font-size: xx-large;
+  margin-top: 5rem;
 }
 
 .body > p {
@@ -123,27 +104,24 @@ h2 {
 
 .highlights {
   display: flex;
-  justify-content: safe;
-  width: fit-content;
+  flex-direction: row;
+  justify-content: space-between;
   margin: 30px;
 }
 
 .highlight {
-  white-space: nowrap;
+  text-wrap: wrap;
   background-color: #d0e1d1;
   border: 3px solid #142b0d;
   border-radius: 8px;
-  padding-bottom: 0px;
-  margin-bottom: 0px;
-  padding: 30px 40px;
+  padding: 40px;
   text-align: center;
-  flex: 1;
   margin: 0 20px;
-  align-content: end;
-}
-
-.highlight:nth-child(2) {
-  padding: 30px;
+  width: 10rem;
+  height: 10rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .highlight > p {
