--- src/styles/LandingPage.css
@@ -1,161 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Itim&display=swap');
-
-.itim-regular {
-  font-family: "Itim", cursive;
-  font-weight: 200;
-  font-style: normal;
-}
-.body {
-  margin: 0;
-  padding: 0;
-  font-family: Helvetica, Sans-Serif;
-}
-
-h1, h2, h3 {
-  font-family: 'Itim';
-}
-
-p { 
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #edece4;
-}
-
-.navbar {
-  position: relative;
-  z-index: 1000;
-  width: 100%;
-  background-color: #3b6838;
-  color: white;
-  display: flex;
-  align-items: center; 
-  padding: 15px 30px;
-}
-
-.title {
-  margin: 0 auto;  
-  font-size: 33px;
-  text-align: center;  
-  position: relative;  
-}
-
-.hamburger {
-  cursor: pointer; 
-  margin: 5px 30px;  
-  padding: 7px;
-  border: 3px solid white;
-  border-radius: 50%;
-  position: absolute;
-  left: 15px; 
-  z-index: 1100; 
-}
-
-.savedCourses {
-  margin: 0px 20px;
-  color: white;
-}
-.savedCourses > a:link {
-  color: white;
-  background-color: transparent;
-  text-decoration: none;
-}
-
-.savedCourses > a:visited {
-  color: white;
-  background-color: transparent;
-  text-decoration: none;
-}
-
-.body {
-  text-align: center;
-  margin: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-}
-
-h2 {
-  color: #3b6838;
-  font-size: xx-large;
-}
-
-.body > p {
-  color: #3b6838;
-  font-size: x-large;
-}
-
-.searchBar {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  background-color: white;
-  width: fit-content;
-  border: 3px solid #8b4513;
-  border-radius: 20px;
-  margin: 30px 20px;
-  padding: 5px;
-}
-
-.searchBar input {
-  width: 400px;
-  margin: 5px 10px;
-  border: none;
-  font-size: medium;
-}
-
-.searchBar input:focus {
-  outline: none;
-  color: #142b0d;
-}
-
-.searchIcon {
-  color: #8b4513;
-  background-color: transparent;
-  text-decoration: none;
-  margin: 10px;
-}
-
-.highlights {
-  display: flex;
-  justify-content: safe;
-  width: fit-content;
-  margin: 30px;
-}
-
-.highlight {
-  white-space: nowrap;
-  background-color: #d0e1d1;
-  border: 3px solid #142b0d;
-  border-radius: 8px;
-  padding-bottom: 0px;
-  margin-bottom: 0px;
-  padding: 30px 40px;
-  text-align: center;
-  flex: 1;
-  margin: 0 20px;
-  align-content: end;
-}
-
-.highlight:nth-child(2) {
-  padding: 30px;
-}
-
-.highlight > p {
-  color: #142b0d;
-}
-
-.highlight a {
-  color: #142b0d;
-  text-decoration: none;
-  font-weight: bold;
-}
-
-.highlight a:hover {
-  color: #3b6838;
-}
